test(fence): move the peer-leg regressions into their own module
The sticky-refusal suite was inserted between the zombie-revoke file's F1 and
F2 sections, pushing that file from 894 to 1218 lines and breaking its own
section order. It is a different property with its own seeding helpers, a
state-log fixture and a deferred-bus adapter test, so it becomes its own
module and the zombie-revoke file returns to exactly what it was.

Two cases reviewers asked for come along: an already-SHARED requester's
re-read leaves a bystander reclaim-zombie refused while capturing for itself,
and a requester joining two SHARED readers is still granted SHARED with
nothing written for either bystander. The claim ladder and the model's
relationship row follow the tests to the new path.
